pyproject.toml
[tool.poetry]
name = "richproject1"
version = "0.1.0"
description = ""
authors = ["frdu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
motor = "^3.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^7.4.4"
uvicorn = "^0.25.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.1.2"
httpx = "^0.26.0"
mongomock-motor = "^0.0.26"
setuptools = "^69.0.3"
python-dotenv = "^1.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
types-pygments = "^2.17.0.20240106"
types-passlib = "^1.7.7.20240106"
types-pyopenssl = "^23.3.0.20240106"
types-python-jose = "^3.3.4.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
